email-service/requirements.txt
Django==5.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
celery==5.4.0
django-celery-results==2.5.1
gunicorn==21.2.0
python-slugify==8.0.1
